+++ backend/src/main/java/com/whereismyhome/amenities/service/AmenService.java
@@ -1,7 +1,9 @@
 package com.whereismyhome.amenities.service;
 
-import com.whereismyhome.amenities.entity.*;
-import com.whereismyhome.amenities.repository.*;
+import com.whereismyhome.amenities.dto.response.BusResponseDto;
+import com.whereismyhome.amenities.repository.BusRepository;
+import com.whereismyhome.houseinfo.dto.HouseResponseDto;
+import com.whereismyhome.houseinfo.repository.HouseInfoRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -10,39 +12,20 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 public class AmenService {
+    private final HouseInfoRepository houseInfoRepository;
     private final BusRepository busRepository;
-    private final HospitalRepository hospitalRepository;
-    private final SchoolRepository schoolRepository;
-    private final SubwayRepository subwayRepository;
-    private final CctvRepository cctvRepository;
 
-    //버스 조회
-    public List<Bus> findBus(){
-        List<Bus> busList = busRepository.findAll();
+    //반경 내 아파트 정보 조회
+    public List<HouseResponseDto> findApt(String lng, String lat) {
+        List<HouseResponseDto> aptList = houseInfoRepository.findByApt(lng, lat);
+
+        return aptList;
+    }
+
+    //반경 내 버스 정보 조회
+    public List<BusResponseDto> findBus(String lng, String lat) {
+        List<BusResponseDto> busList = busRepository.findByBus(lng, lat);
 
         return busList;
-    }
-    //cctv 조회
-    public List<Cctv> findCctv() {
-        List<Cctv> cctvList = cctvRepository.findAll();
-        return cctvList;
-    }
-    //학교 조회
-    public List<School> findSchool(){
-        List<School> schoolList = schoolRepository.findAll();
-
-        return schoolList;
-    }
-    //지하철 조회
-    public List<Subway> findSubway(){
-        List<Subway> subwayList = subwayRepository.findAll();
-
-        return subwayList;
-    }
-    //병원 조회
-    public List<Hospital> findHospital(){
-        List<Hospital> hospitalList = hospitalRepository.findAll();
-
-        return hospitalList;
     }
 }
